requirements.txt
cffi==1.17.1
cryptography==44.0.2
numpy==2.2.5
pillow==11.2.1
pycparser==2.22
scipy==1.15.2
Pillow
numpy
scipy
python-dotenv
cryptography
